Some functions were moved to utils.ts
